dvd: throw out most langpacks
The entire set of language-pack/support is already in dvd-live, and we
do not have the space to include them all over again. Just ship a few
popular ones.

This won't affect people who use the live system for installing, just
the alternate installer. Even then, the missing packages will be
fetched off the net, or through language-selector later on, so it's
currently the least evil removal which will give us the required 300
MB.