make cloud-image a superset of server.
A cloud-image is 'server + cloud things'.

While we're not 100% set on cloud-image being a pure superset of the
'server' esed, the fact is that the live-build based image build process
has made that the case for quite some time (since 12.04 at least).

This removes the need for the cloud image build process to specifically
install the server^ task. 

See bug 1224504 for more information.