separate cloud-image seed from server seed.
Previously, we asserted/assumed that cloud image was superset of server,
but that was not denoted here.  Instead, the cloud image build process
was explicitly installing 'server' as a task to set that.

Now that task installation should not be necessary as the seed reflects it.

After this commit, the following packages that are cloud-image only:
  cloud-guest-utils
  cloud-init
  grub-legacy-ec2
  landscape-client
  libc6-xen
  openssh-server
  overlayroot
  tasksel

server has 'landscsape-common' in it, but that is a dependency of
landscape-client.

After this commit "cloud-image is superset of server".