drop landscape-common from server seed.
There is no reason to be found for this package to be listed in the
server seed, as landscape-common is not much use without landscape-client.
landscape-client is the only rdepends of landscape-common.

landscape-client is still listed in the server-ship seed, so this
should have no affect on the overal support of the package, only
its presense in default 'server' task installs.