server, server-ship: align packages with cloud-image
The over all goal of this commit is to align the server and cloud image
seeds.  The goal is to have "Ubuntu Server" be the same no matter how it
was installed, and ultimately to join these two seeds.

The remaining delta between these two seeds: 
  cloud-init, pollinate, openssh-server

Dropping from server: w3m
Moving from server to server-ship: 
 * wireless-tools
 * wpasupplicant
 * memtest86+

Adding to server things already in cloud image:
  * bcache-tools
  * cloud-guest-utils  # provides growpart
  * cloud-initramfs-copymods
  * cloud-initramfs-dyn-netconf
  * grub-legacy-ec2  # needed if pygrub is loader (Ec2) noninvasive elsewhere
  * ifenslave
  * lvm2
  * lxd
  * mdadm
  * open-iscsi
  * open-vm-tools  # this is inert if not on vmware
  * overlayroot
  * sosreport
  * tmux
  * ubuntu-cloudimg-keyring  # secure dl of ubuntu cloud image
  * vlan
  * wget