live: Explicitly list more live session deps
I think these would get pulled in as dependencies anyway, but it's better to be explicit in case that changes any reason. These are all needed for either auto detecting hardware or auto configuring users/login/etc... during the live session. As they are in the `live` list, they will be removed again in the installed system.